Attempt to introduce parallelism
This doen't yet work. It gave speedups in the NoiseRegimeTest notebook when data was only being read, but for data generation, it gets stuck. It even gets stuck with binomial code data generation, and after a bit of debugging, the issue seems to stem from the recovery matrix taking forever to be found. Since the same recovery can be found when not done in parallel, I think it may be a machine specific issue. I haven't yet tested if random code data generation in parallel works, but I suspect that it wouldn't if the binomial code data generation doesn't. I will try running this on a different machine to see if it works there. If not, I will abandon this approach.